+++ database/migrations/20201031170045_add_parent_thread_to_messages_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddParentThreadToMessagesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->string('parentthread', 128)->nullable();
+            $table->index('thread');
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->dropColumn('parentthread');
+            $table->dropIndex('messages_thread_index');
+        });
+    }
+}